organize to de.metas.handlingunits.allocation.strategy
