+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/allocation/impl/HUProducerDestination.java
@@ -8,6 +8,7 @@ import de.metas.handlingunits.IHandlingUnitsDAO;
 import de.metas.handlingunits.allocation.IAllocationRequest;
 import de.metas.handlingunits.allocation.IAllocationResult;
 import de.metas.handlingunits.allocation.IAllocationStrategy;
+import de.metas.handlingunits.allocation.strategy.AllocationStrategyFactory;
 import de.metas.handlingunits.allocation.transfer.impl.LUTUProducerDestination;
 import de.metas.handlingunits.model.I_M_HU;
 import de.metas.handlingunits.model.I_M_HU_Item;
