+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/allocation/strategy/AllocationStrategyFactory.java
@@ -1,10 +1,11 @@
-package de.metas.handlingunits.allocation.impl;
+package de.metas.handlingunits.allocation.strategy;
 
 import javax.annotation.Nullable;
 
 import org.springframework.stereotype.Service;
 
 import de.metas.handlingunits.allocation.IAllocationStrategy;
+import de.metas.handlingunits.allocation.impl.AllocationDirection;
 import de.metas.quantity.Capacity;
 import lombok.NonNull;
