+++ backend/de.metas.handlingunits.base/src/test/java/de/metas/handlingunits/attributes/impl/split/SplitLUsWithOddQuantitiesTest.java
@@ -45,8 +45,8 @@ import org.w3c.dom.Node;
 import de.metas.business.BusinessTestHelper;
 import de.metas.handlingunits.HUXmlConverter;
 import de.metas.handlingunits.IHandlingUnitsBL;
-import de.metas.handlingunits.allocation.impl.AllocationStrategyFactory;
-import de.metas.handlingunits.allocation.impl.AllocationStrategySupportingServicesFacade;
+import de.metas.handlingunits.allocation.strategy.AllocationStrategyFactory;
+import de.metas.handlingunits.allocation.strategy.AllocationStrategySupportingServicesFacade;
 import de.metas.handlingunits.allocation.transfer.HUTransformService;
 import de.metas.handlingunits.allocation.transfer.HUTransformService.HUsToNewTUsRequest;
 import de.metas.handlingunits.attribute.storage.IAttributeStorage;
